+++ CHANGELOG.md
@@ -89,5 +89,6 @@
 * Fixed doubled submit when using Use key in dialogue
 * Added a brief delay before dialogue exits (effectively debounces Use)
 * Updated Unity Input Mapper config modal message to reflect the new reality
+* Added AnimateQuadScript and TextureAssignScript (hacky/experimental dynamic texture assignment)
 * Implemented dialogue trace (WIP)
 * Implemented lighting for sprite weapon viewmodels (WIP)